componentize globes menu
