+++ src/containers/menus/globes-menu/index.js
@@ -1,0 +1,3 @@
+import Component from './component';
+
+export default Component;
