+++ src/containers/menus/globes-menu/styles.module.scss
@@ -1,0 +1,14 @@
+@import 'styles/settings';
+
+.globesContainer {
+  align-content: end;
+  margin: auto 0 0 0;
+  padding: 0 15px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  width: 100%;
+
+  @media #{$tablet-landscape} {
+    padding: 0 45px;
+  }
+}
